+++ src/main/java/com/example/MnM/boundedContext/board/entity/question/DataNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.MnM.boundedContext.board.question;
+package com.example.MnM.boundedContext.board.entity.question;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
