+++ src/main/java/com/example/MnM/boundedContext/board/entity/question/Question.java
@@ -1,11 +1,13 @@
-package com.example.MnM.boundedContext.board.question;
+package com.example.MnM.boundedContext.board.entity.question;
 
 
-import com.example.MnM.boundedContext.board.answer.Answer;
+import com.example.MnM.boundedContext.board.entity.answer.Answer;
 import jakarta.persistence.*;
 import java.util.List;
 import lombok.ToString;
 import com.example.MnM.boundedContext.board.BaseEntity;
+
+
 import lombok.Getter;
 
 @Getter
