+++ src/main/java/com/example/MnM/boundedContext/board/repository/AnswerRepository.java
@@ -1,13 +1,14 @@
-package com.example.MnM.boundedContext.board.answer;
+package com.example.MnM.boundedContext.board.repository;
 
 
+import com.example.MnM.boundedContext.board.entity.answer.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.transaction.annotation.Transactional;
 
 
-public interface AnswerRepository extends JpaRepository<Answer , Integer> {
+public interface AnswerRepository extends JpaRepository<Answer, Integer> {
     @Transactional
     @Modifying
     @Query(value = "ALTER TABLE answer AUTO_INCREMENT = 1", nativeQuery = true)
